// api/src/controllers/Validators/User/UserControllerValidator.ts
import { z } from "zod";

export const UserControllerValidators = {  
    body: z.object({
        name: z.string().min(2),
        email: z.email(),
        password: z.string().max(8)
    }),

    params: z.object({
        id: z.number()
    }),

    query: z.object({
        limit: z.number(),
        page: z.number()
    }),

    
}